// types/Session.ts
export type LogInType = {
    email?: string
    password?: string
};

export type signUpTp = {
    fullname?: string
    email?: string
    age?: string | number
    password?: string
    emergency_contact?: string | number
    avatar_path?: string
};

export type signUpForm = {
    name?: string
    lastname?: string
    email?: string
    age?: string | number
    password?: string
    emergency_contact?: string | number
    avatar_path?: string
};